Picasa import: Fix integer related security holes in PmpColumnParser
BUG=151701

Review URL: https://chromiumcodereview.appspot.com/14698014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@199854 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/media_galleries/fileapi/picasa/pmp_constants.h
@@ -17,7 +17,7 @@ namespace picasa {
 
 const char* const kPmpExtension = "pmp";
 
-const size_t kPmpHeaderSize = 20;
+const int64 kPmpHeaderSize = 20;
 
 const int kPmpMagic1Offset = 0;
 const int kPmpMagic2Offset = 6;
